create utilities/met_data.py and modify project.py accordingly
